Try setting OLED pixel values to u32 instead of u8 (not tested).

+++ src/piicodev_ssd1306.rs
@@ -127,10 +127,10 @@ impl PiicoDevSSD1306 {
         self.write_cmd(_SET_SEG_REMAP | (rotate & 1))
     }
 
-    pub fn pixel(&mut self, x: u8, y: u8, colour: u8) {
-        fn set_pixel(buffer: &mut [u8], stride: u8, x: u8, y: u8, colour: u8) {
+    pub fn pixel(&mut self, x: u32, y: u32, colour: u8) {
+        fn set_pixel(buffer: &mut [u8], stride: u32, x: u32, y: u32, colour: u8) {
             let index: usize = ((y >> 3) * stride + x) as usize;
-            let offset: u8 = y & 0x07;
+            let offset: u32 = y & 0x07;
 
             let new: u8 = (buffer[index] & !(0x01 << offset)) | ((u8::from(colour != 0)) << offset);
             buffer[index] = new;
@@ -138,7 +138,7 @@ impl PiicoDevSSD1306 {
 
         // let index: usize = (x + y) as usize;
         // self.buffer[index] = colour;
-        set_pixel(&mut self.buffer, WIDTH, x, y, colour)
+        set_pixel(&mut self.buffer, WIDTH as u32, x, y, colour)
     }
 
     pub fn fill_rect(&mut self, x: u8, y: u8, colour: u8) {
